requirements.txt
Django==2.1.2
django-rest-framework==0.1.0
djangorestframework==3.9.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
Unidecode==1.0.22
